conformer_ocr/requirements/requirements.txt
scipy==1.10.0
omegaconf
hydra-core
einops
numpy
matplotlib
imgaug>=0.4.0
tqdm
Pillow
git+https://github.com/parlance/ctcdecode.git
lmdb
timm
torchmetrics
scipy==1.10.0 
evaluate
torch-summary
